shrink Docker image a lot
