+++ scripts/variantstore/wdl/extract/create_vat_inputs_test_files/quickstart_ancestry.tsv
