+++ scripts/variantstore/wdl/extract/requirements.txt
@@ -2,4 +2,3 @@ google-cloud-bigquery
 ijson
 numpy
 google-cloud-storage
-hail==0.2.99
